+++ testoplugin/plugin.go
@@ -15,10 +15,10 @@ package testoplugin
 import "math"
 
 // Priority defines execution order (priority).
-// It defines when a plugin part should be invoked when other parts are available.
+// It defines when a plugin component should be
+// invoked when other parts are available.
 //
-// "Plugin part" means plan, hook, override, etc.
-// Right now, only [Hook]s are supported.
+// "Plugin component" means plan, hook, override, etc.
 type Priority int
 
 const (