+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Support priority for plugin plans and overrides.
+
+### Fixed
+
+- Add missing call to `t.Helper` in runner.
+
 ## [1.3.2] - 2026-05-30
 
 ### Fixed